Create mincost.cpp
Problem: Say there are different stones, where each 'i'th stone has cost Ci and type 'i'. Another functionality exists where for a price of 'k', one type of stone can be converted to another desired type. For given array input of costing of rocks, give the minimum cost to gain all types of stones.

Input:
1(No. of test cases)
3(No. of stones) 10(Price of stone type conversion)
20 10 34(Costs of each stone)

Output:
54

Explanation:
For type3, we buy stone of least cost(type2 of cost=12) and convert it(total cost=12+10=22);
For type1, we must buy stone directly as total cost for conversion is 22 which is greater than cost of the actual stone itself(total cost=22+20=42);
For type2, we buy it directly(Total Cost:=42+12=54)